POM-225 KA5: Fix "back to list" links in offers
Note: this goes back to the proper view but in its initial, empty state.
Ideally, we should get back to the exact same list we were at when clicking
on the offer. Perhaps views should be modified so that the offer displays on top
of the search view (if you go directly via URL, it would just be empty)? That
way, navigating to the list view could preserve what's already there on the
list.