POM-309 Add the JIRA issue collector script to prestage (#289)
Load jQuery and the JIRA issue collector script which requires jQuery but do so
only on the prestage environment.

Also, add the `"esModuleInterop": true` flag to `tsconfig.json` to make it
easier to work with CommonJS-authored modules imported as ES modules. This
allows to import such modules using both `import * as p from 'package'` and
`import p from 'package'`. This is useful in importing jQuery as we need the
default export and without this flag there's a type mismatch.

Co-authored-by: oleksandrdevqube <oleksandr.zderko@devqube.com>